refactor: Changed BarCodeScanner to Camera for Android devices
