+++ packages/mobile/package.json
@@ -20,6 +20,7 @@
         "expo": "38.0.10",
         "expo-asset": "8.2.0",
         "expo-barcode-scanner": "8.2.1",
+        "expo-camera": "8.3.1",
         "expo-cli": "3.25.1",
         "expo-constants": "9.2.0",
         "react": "16.13.1",
