+++ packages/mobile/src/pages/Chat/ConversationList.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { View, Text } from "react-native";
+import { SafeAreaView, Text } from "react-native";
 import styles from "../../styles/chat";
 import MessageItem from "../../components/MessageItem/MessageItem";
 
@@ -7,14 +7,14 @@ const ConversationList: React.FC = () => {
     const [convoPartner, setConvoPartner] = useState(["Jack", "Jill"]);
 
     return (
-        <View style={styles.conversationListContainer}>
-            <View style={styles.header}>
+        <SafeAreaView style={styles.conversationListContainer}>
+            <SafeAreaView style={styles.header}>
                 <Text style={styles.h1}>Moai Chat</Text>
-            </View>
+            </SafeAreaView>
             {convoPartner.map((name, i) => (
                 <MessageItem key={i} guestName={name} />
             ))}
-        </View>
+        </SafeAreaView>
     );
 };
