+++ packages/mobile/src/styles/scanner.ts
@@ -5,7 +5,8 @@ export default StyleSheet.create({
         flex: 1,
         justifyContent: "center",
         alignItems: "center",
-        backgroundColor: "#fff",
+        backgroundColor: "#2d3083",
+        height: "100%",
         paddingTop: Platform.OS === "android" ? StatusBar.currentHeight : 0,
     },
     circleButton: {
@@ -38,4 +39,5 @@ export default StyleSheet.create({
         color: "white",
         fontSize: 25,
     },
+    cameraContainer: {},
 });
